requirements.txt

# Python dependencies for OCR shelf image analysis
pytesseract==0.3.10
pillow==9.5.0
opencv-python==4.8.0.74
fastapi==0.95.2
uvicorn==0.22.0
requests==2.31.0
numpy==1.24.3
pydantic==2.0.3
python-multipart==0.0.6
